requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
openai==1.12.0
openpyxl==3.1.2
tenacity==8.2.3
Flask-Migrate==4.0.5
Flask-Login==0.6.3
gunicorn==21.2.0
